requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-import-export==2.9.0
django-model-history-log==0.2.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
flake8==5.0.4
idna==3.4
import-export==0.2.67.dev6
MarkupPy==1.14
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tablib==3.2.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
